Rename idempotency-spring to idempotency-spring-web
Rename Maven module, directory, and Java package to reflect that this
module only supports Spring Web MVC (not WebFlux).

- Directory: spring/idempotency-spring → spring/idempotency-spring-web
- ArtifactId: idempotency-spring → idempotency-spring-web
- Package: io.github.josipmusa.idempotency.spring → ...spring.web
- Update BOM, starter, and CLAUDE.md references

Closes #39
