+++ spring/idempotency-spring-boot-starter/pom.xml
@@ -22,7 +22,7 @@
         </dependency>
         <dependency>
             <groupId>io.github.josipmusa</groupId>
-            <artifactId>idempotency-spring</artifactId>
+            <artifactId>idempotency-spring-web</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
