+++ spring/idempotency-spring-boot-starter/src/main/java/io/github/josipmusa/idempotency/springboot/IdempotencyAutoConfiguration.java
@@ -3,8 +3,8 @@ package io.github.josipmusa.idempotency.springboot;
 import io.github.josipmusa.core.IdempotencyConfig;
 import io.github.josipmusa.core.IdempotencyEngine;
 import io.github.josipmusa.core.IdempotencyStore;
-import io.github.josipmusa.idempotency.spring.IdempotencyFilter;
-import io.github.josipmusa.idempotency.spring.IdempotentHandlerRegistry;
+import io.github.josipmusa.idempotency.spring.web.IdempotencyFilter;
+import io.github.josipmusa.idempotency.spring.web.IdempotentHandlerRegistry;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.atomic.AtomicInteger;
