+++ spring/idempotency-spring-web/src/main/java/io/github/josipmusa/idempotency/spring/web/Idempotent.java
@@ -1,4 +1,4 @@
-package io.github.josipmusa.idempotency.spring;
+package io.github.josipmusa.idempotency.spring.web;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
