+++ spring/idempotency-spring-web/src/test/java/io/github/josipmusa/idempotency/spring/web/RequestFingerprintTest.java
@@ -1,4 +1,4 @@
-package io.github.josipmusa.idempotency.spring;
+package io.github.josipmusa.idempotency.spring.web;
 
 import static org.assertj.core.api.Assertions.assertThat;
